feat: add message when changing lang and remove using not used
