+++ EasySave/EasySave/Models/LogModel.cs
@@ -1,6 +1,4 @@
 using EasySave.Types;
-using System;
-using System.IO;
 using System.Text.Json;
 using System.Xml.Serialization;
