+++ EasySave/EasySave/ViewModels/DeleteViewModel.cs
@@ -1,10 +1,6 @@
 ﻿using EasySave.Models;
 using EasySave.Views;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+
 using static EasySave.Models.TaskModel;
 
 namespace EasySave.ViewModels
