+++ EasySave/EasySave/ViewModels/HelpViewModel.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using EasySave.Models;
+﻿using EasySave.Models;
 using EasySave.Views;
 
 namespace EasySave.ViewModels
