+++ EasySave/EasySave/Views/CreateView.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Xml.Linq;
-using EasySave.Types;
+﻿using EasySave.Types;
 
 namespace EasySave.Views;
