+++ EasySave/EasySave/Views/ExecuteView.cs
@@ -1,9 +1,4 @@
 ﻿using EasySave.Types;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace EasySave.Views;
