+++ EasySave/EasySave/Views/HelpView.cs
@@ -1,10 +1,5 @@
 ﻿using EasySave.Types;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace EasySave.Views;
