+++ EasySave/EasySave/Views/ModifyView.cs
@@ -1,13 +1,6 @@
 ﻿using EasySave.Types;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace EasySave.Views;
-
-
 
 public class ModifyView : IView
 {
